NXP-22819: move latest created UG page provider impl to core and remove elastic one
