+++ nuxeo-features/nuxeo-elasticsearch/nuxeo-elasticsearch-audit/src/main/resources/META-INF/MANIFEST.MF
@@ -6,5 +6,4 @@ Bundle-Version: 1.0.0
 Nuxeo-Component: OSGI-INF/elasticsearch-audit-contrib.xml,
  OSGI-INF/es-audit-pageprovider-contrib.xml,
  OSGI-INF/admin-pageprovider-contrib.xml,
- OSGI-INF/es-audit-search-contrib.xml,
- OSGI-INF/es-audit-user-contrib.xml
+ OSGI-INF/es-audit-search-contrib.xml
