+++ nuxeo-features/nuxeo-elasticsearch/nuxeo-elasticsearch-audit/src/main/resources/OSGI-INF/es-audit-pageprovider-contrib.xml
@@ -115,10 +115,6 @@
     <maxPageSize>100</maxPageSize>
   </genericPageProvider>
 
-  <genericPageProvider name="LATEST_CREATED_USERS_OR_GROUPS_PROVIDER"
-    class="org.nuxeo.elasticsearch.audit.pageprovider.LatestCreatedUsersOrGroupsPageProvider">
-  </genericPageProvider>
-
   </extension>
 
 </component>
