+++ nuxeo-features/nuxeo-platform-audit/nuxeo-platform-audit-core/src/main/resources/META-INF/MANIFEST.MF
@@ -19,7 +19,7 @@ Import-Package: de.odysseus.el;version="2.1.2",javax.annotation;versio
 Nuxeo-Component: OSGI-INF/nxaudit-service.xml,OSGI-INF/nxaudit-persist
  ence-config.xml,OSGI-INF/core-type-contrib.xml,OSGI-INF/directories-c
  ontrib.xml,OSGI-INF/pageprovider-contrib.xml,OSGI-INF/marshallers-con
- trib.xml
+ trib.xml, OSGI-INF/user-audit-contrib.xml
 Export-Package: org.nuxeo.ecm.platform.audit.listener,org.nuxeo.ecm.pl
  atform.audit.service,org.nuxeo.ecm.platform.audit.service.extension,o
  rg.nuxeo.ecm.platform.audit.service.management
