+++ nuxeo-features/nuxeo-platform-audit/nuxeo-platform-audit-core/src/main/resources/OSGI-INF/user-audit-contrib.xml
